BATTLECLASHPVP.cpp
A two-player battle adventure  game where a player needs to kill another player, both having different statistics/features and also different special abilities, in order to win the game.
The user has the option to select between three players, that is, a Shaolin Master, a Samurai and a Ninja, as the Player 1 and Player 2.
The user has two options to play the game, that is, to attack and to heal, and the user has to combine these options and play in a particular pattern to defeat the enemy player. 
The code helps the user to make a special ability move/attack on the opposite player which will cause more damage than normal damage and reduce a certain amount of health at random occasions.
The game ends when either one of the 2 players die and the victory message for corresponding player will be shown.